# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Foreman AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List companies the current user belongs to
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
        "401":
          description: Unauthorized
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a new company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /companies/{companyId}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get a company by ID
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Updated company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /companies/{companyId}/members:
    get:
      operationId: listCompanyMembers
      tags: [companies]
      summary: List members of a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyMember"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
    post:
      operationId: addCompanyMember
      tags: [companies]
      summary: Add or invite a member to a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyMemberInput"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyMember"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /companies/{companyId}/members/{memberId}:
    patch:
      operationId: updateCompanyMember
      tags: [companies]
      summary: Update a member's role
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyMemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyMember"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: removeCompanyMember
      tags: [companies]
      summary: Remove a member from a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Member removed
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

  # ──── Crews ──────────────────────────────────────────────────────────────
...
  /catalog/materials:
    get:
      operationId: listCatalogMaterials
      tags: [catalog]
      summary: List materials catalog for a company
      parameters:
        - name: companyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Materials catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogMaterial"
        "401":
          description: Unauthorized
    post:
      operationId: createCatalogMaterial
      tags: [catalog]
      summary: Add a material to the catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogMaterialInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogMaterial"
        "401":
          description: Unauthorized

  # ──── Equipment Catalog ───────────────────────────────────────────────────
  /catalog/equipment:
    get:
      operationId: listCatalogEquipment
      tags: [catalog]
      summary: List equipment catalog for a company
      parameters:
        - name: companyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Equipment catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogEquipment"
        "401":
          description: Unauthorized
    post:
      operationId: createCatalogEquipment
      tags: [catalog]
      summary: Add equipment to the catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogEquipmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogEquipment"
        "401":
          description: Unauthorized

  # ──── Voice Dictation Parsing ─────────────────────────────────────────────
...
components:
  schemas:
...
    Company:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    CompanyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
        phone:
          type: string
        email:
          type: string
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
    CompanyMember:
      type: object
      required: [id, companyId, role, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        userId:
          type: ["integer", "null"]
        clerkUserId:
          type: ["string", "null"]
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, supervisor, foreman]
        createdAt:
          type: string
    CompanyMemberInput:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
        role:
          type: string
          enum: [admin, supervisor, foreman]
        firstName:
          type: string
        lastName:
          type: string
    CompanyMemberUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [admin, supervisor, foreman]

    # ── Crew ───────────────────────────────────────────────────────────────
...
    CatalogMaterial:
      type: object
      required: [id, companyId, name, unit, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        unit:
          type: string
        createdAt:
          type: string
    CatalogMaterialInput:
      type: object
      required: [companyId, name, unit]
      properties:
        companyId:
          type: integer
        name:
          type: string
          minLength: 1
        unit:
          type: string
          minLength: 1
    CatalogEquipment:
      type: object
      required: [id, companyId, name, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        type:
          type: ["string", "null"]
        createdAt:
          type: string
    CatalogEquipmentInput:
      type: object
      required: [companyId, name]
      properties:
        companyId:
          type: integer
        name:
          type: string
          minLength: 1
        type:
          type: string

    # ── Dictation ─────────────────────────────────────────────────────────
...