# lib/api-zod/src/generated/types/catalogEquipment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Foreman AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogEquipment {
  id: number;
  companyId: number;
  name: string;
  /** @nullable */
  type?: string | null;
  createdAt: string;
}