# lib/api-zod/src/generated/types/catalogEquipmentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Foreman AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogEquipmentInput {
  companyId: number;
  /** @minLength 1 */
  name: string;
  type?: string;
}