# lib/api-zod/src/generated/types/companyMember.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Foreman AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CompanyMemberRole } from './companyMemberRole';

export interface CompanyMember {
  id: number;
  companyId: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  clerkUserId?: string | null;
  email?: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  role: CompanyMemberRole;
  createdAt: string;
}